refactor(cli): extract core memory verbs into cli_memory.py + fix consolidate shadow
Group the contiguous core CRUD/retrieval commands out of cli.py into
cli_memory.py (save, search, ask, embed, chat-ask, rerank, list, get,
update, reindex, delete, history, ocr-image, provenance, extract-entities,
entities, entity, lint, restore), each a standalone @click.command
registered via add_command. The _resolved AmbiguousIdError wrapper (used
only by these verbs) moves with them. cli.py 4953 -> 4101 LOC.

Also fixes a behavior regression introduced when group registration moved
to the top of cli.py: a long-dead top-level `consolidate` command
(@cli.command def consolidate) had always been shadowed by the
`consolidate` GROUP because the group's decorator ran later in module order.
After centralizing group registration at the top, the later command def
started winning, flipping `memo consolidate` from the group to the dead
one-shot command. The shadowed command is removed entirely, restoring
`memo consolidate` -> group (propose/apply/list-archived).

Verification: ruff clean, mypy clean (87 files), pytest 736 passed,
--help smoke on all 19 verbs + `memo consolidate` resolves to the group.
